Backend/CSSASiteBackend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-rest-knox==5.0.2
djangorestframework==3.16.1
sqlparse==0.5.3
